+++ changelog.md
@@ -1,6 +1,6 @@
 # 4.6.2
 ### Addition
-- Add Phase and Brightness option to chromatic mode
+- Add Reverse, Phase and Brightness option to chromatic mode
 - Collapse mode desc to an info button and remove the ErasLight font to decrease the mod size
 ### Bugfix
 - Fix saturation and duty slider getting toxic in input displayed decimals